portage-hs: Test using pquery output
Add testing of DepBlock to GentooTests, using 'pquery' output to test
DepBlock parsers against.

Currently, it only looks at the packages in 'llvm-core/*::gentoo'. This
is because it is a small subset of packages with a high degree of
complexity in their dependency lists, making them ideal to test
against. Eventually, the plan is to test against all packages in
::gentoo.

Signed-off-by: hololeap <hololeap@protonmail.com>